refactor(ai): move the api/AI shared seams into the domain
Preparation for extracting MCP, chat and the investigation fan-out into their
own Worker. Four modules sat inside the AI surface while services that will
stay in `apps/api` imported them, so the split could not be a clean cut.

Each one moves to where both sides can reach it:

- `chatSessionStub` and the Durable Object's RPC surface become
  `@maple/domain/chat-session-stub`. Its own file rather than the wire
  contract, because addressing the object means handling an unparsed namespace
  handle off a Worker env — the boundary marker belongs on that, not on the
  schemas.
- `widthFor` joins `@maple/domain/investigation-fanout`, beside the payload
  whose `maxWidth` it computes. Its tests come with it.
- `incident-context.ts` moves whole. The staying services and the moving agents
  build the same message, and two copies is how the wording silently drifts.
- `McpToolSurface` joins `@maple/domain/mcp-manifest`, because the audit log
  records it and the audit log is read by code that runs no MCP.

Also deletes the internal Worker-to-Worker RPC surface, which has had no caller
since it was written: `apps/api/src/internal-rpc.ts`, `worker/rpc.ts`, their
wiring, and the `MapleApiRpcContract` half of the domain module. What survived
was never about RPC — a tool's advertised shape and the one failure any surface
can provoke by naming a tool that does not exist — so it becomes
`@maple/domain/mcp-tool-contract`, with `McpToolNotFoundError` tagged
`@maple/mcp/ToolNotFoundError`. Nothing persists that tag on a wire; the only
readers are three `catchTag` call sites in this repo.

The remaining edges from staying code into the moving set are now exactly the
four files that themselves move later: the two runtime graphs and the two chat
route modules.
