+++ apps/api/src/chat/tools.ts
@@ -18,7 +18,8 @@ import { InvestigationId, UserId } from "@maple/domain/primitives"
 import type { RunBudgetHook, RunUsageDelta } from "@effect-agent/engine/RunOptions"
 import { Effect, Option, Schema } from "effect"
 import { Tool, Toolkit } from "effect/unstable/ai"
-import type { McpToolExecutorApi, McpToolSurface } from "@/mcp/dispatcher"
+import type { McpToolExecutorApi } from "@/mcp/dispatcher"
+import type { McpToolSurface } from "@maple/domain/mcp-manifest"
 import { buildMapleToolkit, MapleToolFailure, summarizeToolFailure } from "@/mcp/tools/llm-tools"
 import type { TenantContext } from "@/services/auth/tenant-context"
