+++ apps/api/src/routes/v1/chat-sessions.http.test.ts
@@ -3,7 +3,7 @@ import { OrgId, UserId } from "@maple/domain/http"
 import { WorkerEnvironment } from "@maple/infra/worker-runtime"
 import { ConfigProvider, Effect, Layer, Schema } from "effect"
 import { HttpRouter } from "effect/unstable/http"
-import type { ChatSessionStub } from "@/chat/session"
+import type { ChatSessionStub } from "@maple/domain/chat-session-stub"
 import { Env } from "@/platform/Env"
 import { cleanupTestDbs, createTestDb, type TestDb } from "@/platform/test-pglite"
 import { AuthService } from "@/services/auth/AuthService"
