+++ apps/api/src/services/errors/ai-triage-enqueue.ts
@@ -14,10 +14,10 @@ import { aiTriageSettings, investigations } from "@maple/db"
 import { and, eq, lt } from "drizzle-orm"
 import { Clock, Duration, Effect, Exit, Option, Redacted, Schema } from "effect"
 import { encodeChatTurnTenant } from "@maple/domain/chat-session"
+import { isChatSessionNamespace } from "@maple/domain/chat-session-stub"
 import { Database } from "@/platform/DatabaseLive"
-import { isChatSessionNamespace } from "@/chat/session"
-import { widthFor } from "@/workflows/plan-normalize"
-import { AUTONOMOUS_KICKOFF_LEAD, buildIncidentContextMessage } from "@/workflows/incident-context"
+import { widthFor } from "@maple/domain/investigation-fanout"
+import { AUTONOMOUS_KICKOFF_LEAD, buildIncidentContextMessage } from "@maple/domain/incident-context"
 import { evaluateInvestigationQuota, selectInvestigationUsage } from "@/services/errors/investigation-quota"
 import { startInvestigationFanout } from "@/services/errors/investigation-fanout-start"
 import {
